Generic atomic bit vectors; we compile on a 32-bit platform

+++ examples/bench_elias_fano.rs
@@ -23,7 +23,7 @@ struct Args {
     n: usize,
 
     /// The size of the universe.
-    u: usize,
+    u: u64,
 
     /// The number of values to test.
     t: usize,
